+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG for Sulu Product Bundle
 
 * dev-develop
 
+    * FEATURE     Implemented api for get, adding and delete variant-attribute relations.
     * FEATURE     Implemented new variants api to create, update and delete product variants.
     * FEATURE     Added product type key to id map as parameter to container.
     * BUGFIX      Fixed wrong path to get action schema.